Add passive config option flag to auth command
Adds a "passive" config option and flag to the auth command. When
enabled the certificates are checked for validity and authentication
happens only if one of the certificates has expired.
